Revert "refactor(nuxt): isolate island head via dedicated server plugin"
This reverts commit 2411ac6d22880a9f46f2f48fb5a22a872ee30ec5.
