+++ packages/nuxt/src/app/components/island-renderer.ts
@@ -1,10 +1,15 @@
 import type { defineAsyncComponent } from 'vue'
-import { createVNode, defineComponent, onErrorCaptured } from 'vue'
+import { createVNode, defineComponent, onErrorCaptured, provide } from 'vue'
+import { createHead } from '@unhead/vue/server'
+import { headSymbol } from '@unhead/vue'
 
+import { useNuxtApp } from '../nuxt'
 import { createError } from '../composables/error'
 
 // @ts-expect-error virtual file
 import { islandComponents } from '#build/components.islands.mjs'
+// @ts-expect-error virtual file
+import unheadOptions from '#build/unhead-options.mjs'
 
 export default defineComponent({
   name: 'IslandRenderer',
@@ -15,6 +20,16 @@ export default defineComponent({
     },
   },
   setup (props) {
+    // Use an isolated head for the island so it doesn't include head tags from plugins
+    // and, more importantly, so concurrent route renders don't clobber each other's head
+    // entries via Vue's module-global `currentInstance` (see #32100 regression).
+    const head = createHead(unheadOptions)
+    provide(headSymbol, head)
+    const ssrContext = useNuxtApp().ssrContext
+    if (ssrContext) {
+      ssrContext.head = head
+    }
+
     const component = islandComponents[props.context.name] as ReturnType<typeof defineAsyncComponent>
 
     if (!component) {
